refactor(predicates): renamed the conditionals module to be more clear
