+++ src/prompts/predicates.js
